use pmap instead of @parallel
`@parallel` results in imbalanced work partitions because each iteration may be of different size and the input data may inadvertently be sorted in a manner that penalizes only some workers.

`pmap` schedules each task dynamically and hence will result in a uniform distribution. The additional overhead of pmap can be ignored for large problems, which would be most of the practical workload.

An alternative (not explored here) would be to partition the workload separately and use either `pmap` or `@parallel` just to execute it.